docker/app/requirements.txt
aio-pika==9.3.1
aiormq==6.7.7
idna==3.6
multidict==6.0.4
paho-mqtt==1.6.1
pamqp==3.2.1
psycopg2-binary==2.9.9
pynmeagps==1.0.32
pyrtcm==1.0.14
pyserial==3.5
pyubx2==1.2.35
PyYAML==6.0.1
websockets==12.0
yarl==1.9.4
